Update JdbcConnectionSource.java
If the connection is not closed in the releaseConnection(DatabaseConnection connection) method, then the number of unused (idle) connections will accumulate over time, and exceed the max. simultaneous connections to the DB. E.g. in case of PostgreSQL this exception will come up soon, if you use the close() method only in the original implementation of this class. "FATAL: connection limit exceeded for non-superusers"